发送标签组与中转站的标签组合并去重
